+++ packages/kit/src/core/config/options.js
@@ -123,7 +123,8 @@ const options = object(
 			experimental: object({
 				tracing: object({
 					server: boolean(false)
-				})
+				}),
+				remoteFunctions: boolean(false)
 			}),
 
 			files: object({