+++ packages/kit/src/types/global-private.d.ts
@@ -6,6 +6,8 @@ declare global {
 	const __SVELTEKIT_EMBEDDED__: boolean;
 	/** True if `config.kit.experimental.tracing.server` is `true` */
 	const __SVELTEKIT_SERVER_TRACING_ENABLED__: boolean;
+	/** true if corresponding config option is set to true */
+	const __SVELTEKIT_EXPERIMENTAL__REMOTE_FUNCTIONS__: boolean;
 	/** True if `config.kit.router.resolution === 'client'` */
 	const __SVELTEKIT_CLIENT_ROUTING__: boolean;
 	/**